add missing locale import

+++ src/de/ueller/midlet/gps/GuiOverviewElements.java
@@ -108,7 +108,7 @@ public class GuiOverviewElements extends Form implements CommandListener, ItemSt
 		switch (ovElGroupNr) {
 			case 0:
 				// save overview mode state to node description
-				for (byte i = 1; i < Legend.getMaxType(); i++) {				
+				for (short i = 1; i < Legend.getMaxType(); i++) {				
 					if (Legend.isNodeHideable(i)) {
 						Legend.setNodeOverviewMode(i, ovElSelectionCG.isSelected(count)?overviewMode:nonOverviewMode);
 						count++;
@@ -117,7 +117,7 @@ public class GuiOverviewElements extends Form implements CommandListener, ItemSt
 				break;
 			case 1:
 				// save overview mode state to 'area' description
-				for (byte i = 1; i < Legend.getMaxWayType(); i++) {				
+				for (short i = 1; i < Legend.getMaxWayType(); i++) {				
 					WayDescription w = Legend.getWayDescription(i);
 					if (w.isArea && Legend.isWayHideable(i) ) {
 						Legend.setWayOverviewMode(i, ovElSelectionCG.isSelected(count)?overviewMode:nonOverviewMode);
@@ -127,7 +127,7 @@ public class GuiOverviewElements extends Form implements CommandListener, ItemSt
 				break;
 			case 2:
 				// save overview mode state to way description
-				for (byte i = 1; i < Legend.getMaxWayType(); i++) {				
+				for (short i = 1; i < Legend.getMaxWayType(); i++) {				
 					WayDescription w = Legend.getWayDescription(i);
 					if (!w.isArea && Legend.isWayHideable(i) ) {
 						Legend.setWayOverviewMode(i, ovElSelectionCG.isSelected(count)?overviewMode:nonOverviewMode);
@@ -168,7 +168,7 @@ public class GuiOverviewElements extends Form implements CommandListener, ItemSt
 				frmMaxEl--;
 			}
 
-			byte count=0;
+			short count=0;
 
 			// Warning: do not move this up - it must be after applyElGroupElementStates()
 			ovElGroupNr = (byte) ovElGroupCG.getSelectedIndex();
@@ -196,7 +196,7 @@ public class GuiOverviewElements extends Form implements CommandListener, ItemSt
 			switch (ovElGroupNr) {
 				case 0:
 					// set POI overview states in form				
-					for (byte i = 1; i < Legend.getMaxType(); i++) {				
+					for (short i = 1; i < Legend.getMaxType(); i++) {				
 						if (Legend.isNodeHideable(i)) {
 							ovElSelectionCG.append(Legend.getNodeTypeDesc(i), Legend.getNodeSearchImage(i));
 							ovElSelectionCG.setSelectedIndex(count, ((Legend.getNodeOverviewMode(i) & Legend.OM_MODE_MASK) == Legend.OM_OVERVIEW) );
@@ -206,7 +206,7 @@ public class GuiOverviewElements extends Form implements CommandListener, ItemSt
 					break;
 				case 1:
 					// set Area overview states in form
-					for (byte i = 1; i < Legend.getMaxWayType(); i++) {				
+					for (short i = 1; i < Legend.getMaxWayType(); i++) {				
 						WayDescription w = Legend.getWayDescription(i);
 						if (w.isArea && Legend.isWayHideable(i) ) {
 							ovElSelectionCG.append(w.description, areaImage(w.lineColor));
@@ -217,7 +217,7 @@ public class GuiOverviewElements extends Form implements CommandListener, ItemSt
 					break;
 				case 2:
 					// set Way overview  states in form
-					for (byte i = 1; i < Legend.getMaxWayType(); i++) {				
+					for (short i = 1; i < Legend.getMaxWayType(); i++) {				
 						WayDescription w = Legend.getWayDescription(i);
 						if (!w.isArea && Legend.isWayHideable(i) ) {
 							ovElSelectionCG.append(w.description, wayImage(w));
